mmcy Audit/mmcyAudit/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.6
django-active-link==0.1.8
django-appconf==1.0.5
django-avatar==7.0.1
django-import-export==3.0.2
django-jazzmin==2.6.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
Pillow==9.4.0
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
